requirements.txt
pandas==1.3.5
python-dotenv==0.21.0
requests==2.27.1
PyTikTokApi==0.0.5
youtube_dl==2021.12.17
